+++ packages/core/src/scales/updates.ts
@@ -1,6 +1,7 @@
 import { AxisScale, AxisScaleProps, NumericScaleProps } from './types'
 import { cloneDeep } from 'lodash'
 import { createContinuousScale } from './axes'
+import { getMinMax } from './utils'
 
 /** create scale props with a size key */
 export const fillSize = (props: Omit<AxisScaleProps, 'size'>, size: number) => {
@@ -21,6 +22,13 @@ export const zoomDomain = (
     const rangeSize = range[1] - range[0]
     const margin = (rangeSize * 0.5) / zoomFactor
     result.viewDomain = [scale.invert(center - margin), scale.invert(center + margin)]
+    return result
+}
+
+/** create scale props with a changed domain (target are coordinates to be converted into a new domain) */
+export const changeDomain = (props: AxisScaleProps, scale: AxisScale, target: [number, number]) => {
+    const result = cloneDeep(props)
+    result.viewDomain = getMinMax([scale.invert(target[0]), scale.invert(target[1])])
     return result
 }