+++ packages/core/tests/interactivity/Button.test.tsx
@@ -27,4 +27,17 @@ describe('Button', () => {
         expect(button.getAttribute('class')).toContain('button abc')
         expect(button.querySelector('rect')).toBeDefined()
     })
+
+    it('creates a button with selected state', () => {
+        render(
+            <Chart {...chartProps}>
+                <Button variant={'A'} position={[20, 20]} selected />
+                <Button variant={'B'} position={[20, 20]} selected className={'custom'} />
+            </Chart>
+        )
+        const buttonA = screen.getByRole('button-A')
+        expect(buttonA.getAttribute('class')).toContain('button selected A')
+        const buttonB = screen.getByRole('button-B')
+        expect(buttonB.getAttribute('class')).toContain('button selected B')
+    })
 })